backend/ai-service/requirements.txt
Flask==2.3.0
Flask-CORS==4.0.0
numpy==1.24.3
opencv-python==4.8.0.74
librosa==0.10.0
soundfile==0.12.1
tensorflow==2.13.0
scikit-learn==1.3.0
joblib==1.3.1
Pillow==10.0.0
